feat: add shared kv utilities
